Refactor: recoil atom을 별도 recoil 디렉터리에서 관리
