+++ src/hooks/recoil/atom.ts
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+import { UserType } from '../../types';
+
+export const userState = atom<UserType>({
+  key: 'userState',
+  default: {},
+});
