+++ src/utils/localStorage.ts
@@ -1,10 +1,7 @@
 export const LocalStorage = (function () {
   const setAuthFlag = () => localStorage.setItem('auth', 'auth');
 
-  const removeAuthFlag = () => {
-    console.log('remove user');
-    localStorage.removeItem('auth');
-  };
+  const removeAuthFlag = () => localStorage.removeItem('auth');
 
   const validateAuthFlag = () => {
     const authFlag = localStorage.getItem('auth');
